requirements.txt
flask
onnxruntime
Pillow
